chore: move nav to wsId root dir
