+++ apps/web/src/app/[locale]/(dashboard)/[wsId]/nav.tsx
